Remove deprecated Xorg classes
