+++ NativeLib/Linux/Xorg/XRRMonitorInfo.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Runtime.InteropServices;
 
-namespace NativeLib.Linux
+namespace NativeLib.Linux.Xorg
 {
     [StructLayout(LayoutKind.Sequential)]
     public struct XRRMonitorInfo
